+++ CHANGELOG.md
@@ -18,6 +18,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **Other changes:**
 
 * **SOUP can now be installed as a Chrome extension.** (Thanks to Peanut for help and for prodding me into actually doing this!)
+* Use a less hacky way to obtain the site and question IDs for realtime update subscription. (Thanks, enki-code!)
 * Improved [mse259692](http://meta.stackexchange.com/q/259692) fix by hacking the JS Date constructor and methods to pretend that local time is UTC.
 * Tweaked [mse217779](http://meta.stackexchange.com/q/217779) CSS to slightly slow down spoiler fade-in per [request](http://meta.stackexchange.com/q/278935) and to add a notice text to spoilers per [another request](http://meta.stackexchange.com/q/104085).
 * Rewrote the [math19650](http://meta.math.stackexchange.com/q/19650) fix to handle any display math syntax (including e.g. `\begin{equation}` ... `\end{equation}`) and to selectively disable display math in elements that should not contain it (currently comments, post summaries, question titles and the sidebar).